ready to present

+++ src/main/java/com/epam/capstone/controllers/ProductController.java
@@ -7,29 +7,26 @@ import com.epam.capstone.entities.Product;
 import com.epam.capstone.security.CustomUserDetails;
 import com.epam.capstone.services.ProductServiceImpl;
 import jakarta.validation.Valid;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.Banner;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.logging.Logger;
 
 @Controller
 public class ProductController {
     private final ProductServiceImpl productService;
 
+    private static final Logger logger = Logger.getLogger(ProductController.class.getName());
 
     public ProductController(ProductServiceImpl productService) {
         this.productService = productService;
@@ -44,6 +41,24 @@ public class ProductController {
         List<ProductBasicDto> products= productService.getAllProducts(pageable);
         model.addAttribute("products",products);
         return "products";
+    }
+
+
+    @PreAuthorize("hasRole('USER')")
+    @GetMapping(value = "/products/myproducts")
+    public String getMyProducts(Model model){
+        logger.info("in method");
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication == null || !(authentication.getPrincipal() instanceof CustomUserDetails)) {
+            return "redirect:/login";
+        }
+        logger.info("authenticatio is not null");
+        CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+        logger.info("casted successfully");
+        List<ProductDto> productList = productService.getProductsDtoBySellerId(userDetails.getUserId());
+        logger.info("product list:" + productList.toString());
+        model.addAttribute("productList",productList);
+        return "myproducts";
     }
 
     @GetMapping(value = "/products/id/{product_id}")
@@ -81,14 +96,17 @@ public class ProductController {
 
 
     @GetMapping(value = "/products/seller/{seller_username}")
-    public List<ProductBasicDto> getProductsBySellerUsername(
+    public String getProductsBySellerUsername(
+            Model model,
             @PathVariable String seller_username,
             @RequestParam(name = "page", defaultValue = "0") int page,
             @RequestParam(name = "size", defaultValue = "20") int size,
             @RequestParam(name = "sort", defaultValue = "rating") String sortField
     ) {
         PageRequest pageable = PageRequest.of(page, size, Sort.by(sortField).descending());
-        return productService.getProductsBySellerUsername(seller_username, pageable);
+        List<ProductBasicDto> products= productService.getProductsBySellerUsername(seller_username, pageable);
+        model.addAttribute("products",products);
+        return "sellerproducts";
     }
 
     @GetMapping(value = "/products/search")
@@ -107,17 +125,6 @@ public class ProductController {
         return "products";
     }
 
-//    @ResponseBody
-//    @GetMapping(value = "/products/search/{name}")
-//    public List<ProductBasicDto> getSearchResult(
-//
-//            @PathVariable String name,
-//            @RequestParam(name = "page", defaultValue = "0") int page,
-//            @RequestParam(name = "page", defaultValue = "12") int size
-//    ) {
-//        PageRequest pageable = PageRequest.of(page, size);
-//        return productService.getSearchResult(name, pageable);
-//    }
 
     @GetMapping(value = "/products/filter/")
     public String  getFilteredAndSortedProducts(
@@ -138,44 +145,51 @@ public class ProductController {
 
 
     @PreAuthorize("hasRole('USER')")
-    @DeleteMapping(value = "/products/delete/{productId}")
-    public ResponseEntity<String> deleteProduct(@PathVariable Integer productId) {
-        Product product = productService.getProductById(productId);
-
+    @GetMapping(value = "/products/delete/{productId}")
+    public String deleteProduct(@PathVariable Integer productId,RedirectAttributes redirectAttributes){
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication == null || !(authentication.getPrincipal() instanceof CustomUserDetails)) {
+            return "redirect:/login";
+        }
         CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
 
+        Product product = productService.getProductById(productId);
         if (product.getSeller().getId().equals(userDetails.getUserId())) {
             productService.deleteProduct(productId);
-            return ResponseEntity.ok("Product deleted successfully.");
+            redirectAttributes.addFlashAttribute("successMessage","Product deleted successfully!");
+            return "redirect:/products/myproducts";
         } else {
-            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("You don't have permission to delete this product.");
+            redirectAttributes.addFlashAttribute("failureMessage","You don't have permission to delete this product");
+            return "redirect:/forbidden";
         }
+
     }
+
+
+
     @PreAuthorize("hasRole('USER')")
-    @GetMapping(value = "/product/add")
+    @GetMapping(value = "/products/add")
     public String productAddForm(Model model){
         model.addAttribute("product",new ProductPlacingDto());
-        return "placeproduct";
+        return "addproduct";
     }
 
     @PreAuthorize("hasRole('USER')")
-    @PostMapping("/products/add")
+    @PostMapping("/products/add/save")
     public String addProduct(@Valid @ModelAttribute("product") ProductPlacingDto product,
                                              BindingResult result,
-                                             Model model,
                                              RedirectAttributes redirectAttributes
     ) {
         try {
             if (result.hasErrors()) {
-                return "placeproduct";
+                return "redirect:/products/add";
             }
             productService.saveProduct(product);
             redirectAttributes.addFlashAttribute("successMessage","Product added successfully!");
-            return "profile";
+            return "redirect:/products/myproducts";
         }catch (Exception e){
-            model.addAttribute("registrationError", "An error occurred during registration. Please try again.");
-            return "placeproduct";
+            redirectAttributes.addFlashAttribute("errorMessage", "An error occurred during registration. Please try again.");
+            return "redirect:/products/add";
         }
     }
 }
